+++ src/js/components/ButtonArea.css
@@ -1,0 +1,8 @@
+.button-area {
+  margin: 10px 15px;
+}
+
+.button {
+  display: inline-block;
+  margin: 10px 5px;
+}
