+++ src/js/components/Contents.css
@@ -3,35 +3,3 @@
   /* background-color: #E0FFFF; */
   text-align: center;
 }
-
-.button-area {
-  margin: 10px 15px;
-}
-
-.button {
-  display: inline-block;
-  margin: 10px 5px;
-}
-
-.text-area {
-  width: 80%;
-  margin: 20px auto 0;
-}
-.speech-text {
-  width: 100%;
-  height: 100px;
-}
-
-.select-area {
-  width: 80%;
-  margin: 20px auto 0;
-}
-
-.state-area {
-  width: 80%;
-  margin: 20px auto 0;
-}
-
-.state-area img {
-  width: 400px;
-}
