+++ src/js/components/Contents.js
@@ -1,102 +1,27 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { RaisedButton, SelectField, MenuItem } from 'material-ui';
-import style from './Contents.css';
 import CSSModules from 'react-css-modules';
-import { APIS } from '../constants';
-import { startRecognition, stopRecognition } from '../actions/recognition';
-import { startSynthesis } from '../actions/synthesis';
-import { changeText } from '../actions/text';
-import { setApi } from '../actions/api';
+import style from './Contents.css';
+import SelectArea from './ButtonArea';
+import ButtonArea from './ButtonArea';
+import TextArea from './TextArea';
+import StatusArea from './StatusArea';
 
-class Contents extends React.Component {
+@CSSModules( style )
+export default class Contents extends React.Component {
 
   constructor( props ) {
     super( props );
   }
 
-  start() {
-    this.props.startRecognition();
-  }
-
-  stop() {
-    this.props.stopRecognition();
-  }
-
-  setText( e ) {
-    this.props.changeText( e.target.value );
-  }
-
-  changeRecognitionApi( e, idx, val ) {
-    this.props.setApi( val, 'RECOGNITION' );
-  }
-
-  changeSynthesisApi( e, idx, val ) {
-    this.props.setApi( val, 'SYNTHESIS' );
-  }
-
   render() {
-    var recognitionApis = Object.keys( APIS.RECOGNITION ).map(( k, i ) => {
-      return (
-        <MenuItem key={ 'recognition-menu-item' + i } value={ k } primaryText={ k } />
-      );
-    });
-    var synthesisApis = Object.keys( APIS.SYNTHESIS ).map(( k, i ) => {
-      return (
-        <MenuItem key={ 'synthesis-menu-item' + i } value={ k } primaryText={ k } />
-      );
-    });
-
-    var imageFiles = {
-      NORMAL: 'NORMAL.png',
-      LISTENING: 'LISTENING.jpg',
-      SPEAKING: 'SPEAKING.jpg',
-    };
-
     return(
       <div styleName="section">
-        <div styleName="select-area">
-          <SelectField
-            value={ this.props.api.RECOGNITION }
-            onChange={ ::this.changeRecognitionApi }
-            floatingLabelText="Recognition"
-          >
-            { recognitionApis }
-          </SelectField>
-          <SelectField
-            value={ this.props.api.SYNTHESIS }
-            onChange={ ::this.changeSynthesisApi }
-            floatingLabelText="Synthesis"
-          >
-            { synthesisApis }
-          </SelectField>
-        </div>
-        <div styleName="button-area">
-          <RaisedButton styleName="button" onClick={ ::this.start } label="はじめる" primary={true} />
-          <RaisedButton styleName="button" onClick={ ::this.stop } label="やめる" secondary={true} />
-        </div>
-        <div styleName="text-area">
-          <textarea styleName="speech-text" value={ this.props.text } onChange={ ::this.setText }></textarea>
-        </div>
-        <div styleName="state-area">
-          <img src={ './assets/images/' + imageFiles[this.props.status] } />
-        </div>
+        <SelectArea />
+        <ButtonArea />
+        <TextArea />
+        <StatusArea />
       </div>
     );
   }
 }
-
-export default connect( state => ({
-  recognition: state.recognition,
-  synthesis: state.synthesis,
-  text: state.text,
-  api: state.api,
-  status: state.status
-}), {
-  startRecognition,
-  stopRecognition,
-  startSynthesis,
-  changeText,
-  setApi
-})( CSSModules( Contents, style ));
