+++ src/js/components/SelectArea.css
@@ -1,0 +1,4 @@
+.select-area {
+  width: 80%;
+  margin: 20px auto 0;
+}
