+++ src/js/components/StatusArea.css
@@ -1,0 +1,8 @@
+.state-area {
+  width: 80%;
+  margin: 20px auto 0;
+}
+
+.state-area img {
+  width: 400px;
+}
