+++ src/js/components/TextArea.css
@@ -1,0 +1,8 @@
+.text-area {
+  width: 80%;
+  margin: 20px auto 0;
+}
+.speech-text {
+  width: 100%;
+  height: 100px;
+}
